+++ src/wrapper/jnienv.rs
@@ -1982,6 +1982,10 @@ impl<'a> JNIEnv<'a> {
     /// release_byte_array_elements() is called.
     ///
     /// See also [`release_byte_array_elements`](struct.JNIEnv.html#method.release_byte_array_elements)
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_byte_array_elements instead"
+    )]
     pub fn get_byte_array_elements(&self, array: jbyteArray) -> Result<(*mut jbyte, bool)> {
         non_null!(array, "get_byte_array_elements array argument");
         let mut is_copy: jboolean = 0xff;
@@ -2008,6 +2012,10 @@ impl<'a> JNIEnv<'a> {
     /// control over memory management, and should be used with extreme care.
     ///
     /// See also [`commit_byte_array_elements`](struct.JNIEnv.html#method.commit_byte_array_elements)
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_byte_array_elements instead"
+    )]
     pub fn release_byte_array_elements(
         &self,
         array: jbyteArray,
@@ -2029,6 +2037,10 @@ impl<'a> JNIEnv<'a> {
     ///
     /// This function has no effect if elems is not a copy of the elements in array. Otherwise,
     /// this function copies back the content of the array (and does not free the elems buffer).
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_byte_array_elements instead"
+    )]
     pub fn commit_byte_array_elements(&self, array: jbyteArray, elems: &mut jbyte) -> Result<()> {
         non_null!(array, "commit_byte_array_elements array argument");
         jni_void_call!(
@@ -2058,8 +2070,10 @@ impl<'a> JNIEnv<'a> {
         array: jbyteArray,
         mode: ReleaseMode,
     ) -> Result<AutoByteArray> {
-        let (ptr, is_copy) = self.get_byte_array_elements(array)?;
-        AutoByteArray::new(self, array.into(), ptr, mode, is_copy)
+        non_null!(array, "get_auto_byte_array_elements array argument");
+        let mut is_copy: jboolean = 0xff;
+        let ptr = jni_non_void_call!(self.internal, GetByteArrayElements, array, &mut is_copy);
+        AutoByteArray::new(self, array.into(), ptr, mode, is_copy == sys::JNI_TRUE)
     }
 
     /// Return a tuple with a pointer to elements of the given Java primitive array as first
@@ -2088,6 +2102,10 @@ impl<'a> JNIEnv<'a> {
     ///
     /// See also [`get_byte_array_elements`](struct.JNIEnv.html#method.get_byte_array_elements)
     /// See also [`release_primitive_array_critical`](struct.JNIEnv.html#method.release_primitive_array_critical)
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_primitive_array_critical instead"
+    )]
     pub fn get_primitive_array_critical(&self, array: jarray) -> Result<(*mut c_void, bool)> {
         non_null!(array, "get_primitive_array_critical array argument");
         let mut is_copy: jboolean = 0xff;
@@ -2107,6 +2125,10 @@ impl<'a> JNIEnv<'a> {
     ///
     /// See also [`get_primitive_array_critical`](struct.JNIEnv.html#method.get_primitive_array_critical)
     /// See also [`commit_primitive_array_critical`](struct.JNIEnv.html#method.commit_primitive_array_critical)
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_primitive_array_critical instead"
+    )]
     pub fn release_primitive_array_critical(
         &self,
         array: jarray,
@@ -2128,6 +2150,10 @@ impl<'a> JNIEnv<'a> {
     ///
     /// This function has no effect if elems is not a copy of the elements in array. Otherwise,
     /// this function copies back the content of the array (and does not free the elems buffer).
+    #[deprecated(
+        since = "0.19.0",
+        note = "Please use get_auto_primitive_array_critical instead"
+    )]
     pub fn commit_primitive_array_critical(
         &self,
         array: jbyteArray,
@@ -2170,8 +2196,16 @@ impl<'a> JNIEnv<'a> {
         array: jarray,
         mode: ReleaseMode,
     ) -> Result<AutoPrimitiveArray> {
-        let (ptr, is_copy) = self.get_primitive_array_critical(array)?;
-        AutoPrimitiveArray::new(self, array.into(), ptr, mode, is_copy)
+        non_null!(array, "get_auto_primitive_array_critical array argument");
+        let mut is_copy: jboolean = 0xff;
+        let ptr = jni_non_void_call!(
+            self.internal,
+            GetPrimitiveArrayCritical,
+            array,
+            &mut is_copy
+        );
+        non_null!(ptr, "get_primitive_array_critical return value");
+        AutoPrimitiveArray::new(self, array.into(), ptr, mode, is_copy == sys::JNI_TRUE)
     }
 }